feat(progression): add weapon/passive leveling, evolution gating, and upgrade picker overhaul
Weapons and passives now scale L1-L5 (damage, fire rate, projectile speed,
radius/orbit count); evolutions require both prerequisites at max level.
Adds upgrade_applied event to fix picker getting stuck on auto-select
timeout, fixes game freeze when the upgrade pool is empty, and reworks
the UpgradePicker overlay with a countdown timer, loadout sidebar, and
stat-delta display.

+++ apps/game-server/src/schema/GameSchema.ts
@@ -19,8 +19,8 @@ export class PlayerSchema extends Schema {
   @type('uint32') declare y: number
   @type('uint8') declare hp: number // max 255 — player max HP fits
   @type('uint8') declare maxHp: number
-  @type('uint8') declare level: number // max 255 — level cap fits
-  @type('uint8') declare xp: number // xp within current level (resets at threshold)
+  @type('uint32') declare level: number
+  @type('uint32') declare xp: number
   @type(['string']) weapons = new ArraySchema<string>()
   @type(['string']) passives = new ArraySchema<string>()
 }
